feat: 로그아웃시 fcmToken 처리 추가

+++ src/main/java/com/ssu/ongi/domain/auth/service/AuthCommandService.java
@@ -58,9 +58,13 @@ public class AuthCommandService {
         return LoginResponse.of(tokens.accessToken(), tokens.refreshToken(), request.loginMode(), member);
     }
 
-    public void logout(Long memberId) {
+    public void logout(Long memberId, LoginMode loginMode) {
+        if (loginMode == LoginMode.ELDER) {
+            throw new GeneralException(ErrorStatus.ELDER_CANNOT_LOGOUT);
+        }
         tokenCommandService.logout(memberId);
         memberCommandService.deleteFcmToken(memberId);
+        elderCommandService.deleteFcmToken(memberId);
     }
 
     public void withdraw(Long memberId) {
